Overhaul use of external commands
Incorporating lessons learned from the writing of the floxadm shell
script. In that case we were running from the UNCLE and in that context
it was important to differentiate between the commands we expected to
find in the UNCLE and those we were required to run from the base O/S
(anything setuid, for example). We cannot simply wrap and prepend to
the PATH as with other Nix scripts because we don't want these UNCLE
paths to leak out into subcommands (which won't be able to access the
UNCLE's mount namespace).

Also refactored the Makefile and Nix expression to perform "@@"
replacements from the Makefile only.