refined the invoke() and command hashing in support of --verbose
The goal of the --verbose flag is to print out the pertinent commands
being invoked, which requires wrapping those invocations witn the
invoke() function. However, we don't want to wrap all invocations,
but just those that directly relate to the results/output, so this
patch introduces the following convention:

- $_commandname = path to command
- $invoke_commandname = "invoke $_commandname"

In this way we can call $_foo when we want to run something, or
$invoke_foo when we want that to print out a verbose message when
running with the --verbose flag.

N.B. one cannot invoke "$_xargs $invoke_foo" because xargs will not
be able to find the "invoke" command. One must always use the $_foo
argument with xargs.