Add NIX_SSL_CERT_FILE to environment
The RPM installer on Amazon Linux highlighted the need to include
the cacert bundle with the RPM, and this change ensures that flox
uses that bundle for all invocations of Nix. In truth this is not
the best solution, and that we should instead be curating our own
flox-managed profile of "system" packages, but this will do for
now as a bootstrap mechanism.

Also export NIX_REMOTE=daemon as a flox default.