updating catalogAttrPathPrefix
This needs to change in response to changes on the capacitor side.

Background: the "legacyPackages" flake namespace is special-cased in
the Nix code to allow recursive descents for the purposes of searching,
and while it's possible to get this behavior in other parts of the
namespace by defining the recurseIntoAttrs attribute at each level
of the catalog hierarchy that is somewhat onerous and makes the code
messy, so Tom instead moved the catalog under the legacyPackages
namespace.