default.nix: wrap `gh` (take two), and `nix` too!
Some programs cannot function without git, ssh, and other
programs in their PATH. We have gone gone to great lengths
to avoid leaking /nix/store paths into PATH, so in order
to correct for these shortcomings we need to arrange for
flox to invoke our wrapped versions of these programs in
preference to the ones straight from nixpkgs.

This replaces the previous commit, which worked fine but
also served to reveal that `nix` itself suffered from the
same problem and inspired this more scalable approach.