Wrap flox command to set env variables required by bash
As documented in https://github.com/flox/flox/issues/45 the bash
executable itself will emit locale warning messages if invoked
a) with $LC_ALL set and b) without a set of working locales.

This patch adds a new C wrapper to be invoked before launching
the flox (bash) script, and updates the installation to install
this wrapper to the bin directory while we move the actual flox.sh
script into the libexec/flox directory.