Make handling of exported variables consistent
pkgs.glibcLocales is defined but cannot be evaluated on Darwin, so
strictly speaking we shouldn't be looking to import it in the first
place. This highlighted that we should be consistent with our handling
of package references within the lib.optional* conditional blocks.
This patch makes this handling more regular and updates the C wrapper
to similarly export the SSL_CERT_FILE environment variable.