Don't rebuild links/profiles for generations that already exist
Whenever anything changed in a profile flox was being super diligent
in verifying the contents of each generation. We can be more efficient
by trusting that if the link and the directory to which it points
exists then it's likely to be fine as-is.

Also refactor/cleanup slightly with $targetLink variable. This diff
is very simple when viewed with `git show -b`.