Add support for `flox edit` (#68)
The origins for the flox edit feature are discussed in issue#62, but
in short:

* we want to offer the ability to manage the contents of profiles
  in a declarative manner
* we need a place to master this declarative "manifest"
* the <organization>/floxmeta repository provides a reasonable
  structure for storing this manifest (see issue#14 for more details)

This patch adds support for a new manifest.toml for storing the
declarative manifest for a profile. It also prompts the user to
confirm when switching a profile between imperative and
declarative configuration modes. In future we may drop the
notion of configuration modes and provide both imperative
and declarative ways of editing the same profiles.

The new profile logic was correctly creating an orphan branch but
forgetting to ensure the new branch started empty. Also clean up
logic for bootstrapping profiles using declarative method.