Eliminate profile modes, update `flox history` semantics
With this commit we eliminate the notion of distinct imperative and
declarative profile modes, allowing the (declarative) management of
profiles with *any* of `flox {install,remove,update,edit}`. This is
what most people want, the ability to manage profiles in whatever
way makes the most sense at the time.

Notably, this feature preserves comments and other sections within
the TOML configuration file, but at the cost of imposing the constraint
that the "[packages]" section must appear at the end of the file. I'm
sure we can fix this with a more advanced TOML parser but I'm punting
on this requirement as I only have dasel to work with, and it should be
much easier once we port this code to golang.

Also updated the `flox history` command to make the previous `--verbose`
mode the default and to accept a `--oneline` argument like `git log`.