`flox edit`: retry edit with invalid TOML
It's annoying to make a bunch of changes to the TOML manifest and
then lose those edits with a single typo. This update makes the
`flox edit` function repeat the editor invocation on the temporary
file for as long as it takes for the user to produce valid TOML or
give up trying.