Overhaul configuration / github bootstrapping (#87)
This commit represents an almost-complete overhaul of the process of
bootstrapping the flox user git-related configuration. There's been
so much refactoring that the diff is hard to parse so I've summarized
the changes below.

It was observed that the git initial user dialog had the potential to
emit warnings at any point, which had the potential to confuse users
trying to follow our tutorials. This commit adds the checkGitConfig()
function and calls it from the flox bootstrap to validate the user's
git configuration settings. See:
    https://github.com/flox/flox/issues/75#issuecomment-1162971101

We also observed that it didn't make sense to prompt the user to confirm
their github username when the process of logging in with the `gh` CLI
identified the username at the same time. This patch changes the process
of initializing the flox git-related user variables as follows:

* bootstrap main()
  * if in interactive mode
    * set gitBaseURL, defaultFlake, organization from defaults
    * verify github login with `gh auth status`, else `gh auth login`
    * set username from `gh` client data
  * else
    * set gitBaseURL, defaultFlake, organization from defaults
    * set username from `gh` client data if it exists, else $USER

Also created new parseURL() and overhauled gitBaseURLToFlakeURL() in
support of the above.

Updated the `git` invocation within metaGit() to use the `gh` credential
helper for github.com specifically. Note that the `git` invocations that
happen via `nix` already benefit from the `gh` token by virtue of the
`access-tokens` variable in `nix.conf`.

Final change was to update the default gitBaseURL from "git+ssh@github.com:"
to "https://github.com/". I'm happy doing this for two reasons:

1. We must do this for the public PLG release eventually in any case, once
   our root flake and capacitor have been marked as public repositories.
2. With this commit we're taking responsibility for shepherding the user
   through the process of logging into github.com using the `gh` CLI, which
   makes it possible to depend on the availability of a valid oauth token.
3. We're planning to glean inspiration from the `gh` CLI when we rewrite
   the `flox` CLI in golang, so it makes sense to start leveraging the `gh`
   authentication methods now in anticipation of that effort.