silence dasel complaint reading TOML config
When looking for an attribute not found within the TOML configuration
dasel was emitting a warning message. This commit rewrites the logic
so that we instead convert the entire TOML to JSON and then use jq
to more carefully parse the results, taking care to return zero when
the requested data is not found.

Also replace the previous output processing with a single eval() to
avoid double-quoting woes.