hotfix to commit a75e40d4
The previous dasel patch contained two mistakes:
- forgot that the results of read_flox_conf() were already being eval'd
  from the calling process, so there's no need to eval within the function
- forgot the "$var" part of the "FLOX_CONF_\($var\)_\(.key)" variable name

This bug also revealed a bootstrapping problem when the contents of the
floxUserMeta.json is "corrupt", or in this case contains a gitBaseURL with
an empty value. In this situation even a "flox config --reset" will not
help, and the only workaround is to "rm ~/.config/flox/floxUserMeta.json".
We will ultimately be fixing this by restructuring the code to avoid reading
this configuration file until the contents are needed, but I'm including
a hack/workaround for this specific error in the registry get() function
to prevent people from having to manually remove that file if and when
corruption does happen.