tuple reversal fixes
The previous commit changed the floxpkgs tuple from
channel.stability.attrPath to stability.channel.attrPath,
and this change needs to be reflected in the output of `flox list`
and for the parameters accepted by `flox install`. This patch
addresses both of these.

This patch also bumps the version to reflect the significant
and user-facing impact of the tuple reordering.