refactor: rename gitBaseURL -> git_base_url w/ transition logic (#502)
https://github.com/flox/flox-private/pull/288 documents the effort to
implement hierarchical configuration files and the library we want to
use requires the use of snake case rather than camel case. This patch
updates the bash version to a) parse the snake case value from the new
location of the user-specific config file, b) update the default
etc/flox.toml configuration file to use the snake case, and c) provides
a default value if all parsing goes wrong. With this we should always
have a value for $git_base_url and be protected through the transition
on the bash side.

This patch also renames $gitBaseURL to $git_base_url throughout.