fix: don't copy files from flox-floxpkgs (#517)
Copy files mastered in ${inputs.flox-floxpkgs}/lib back into the
repository. We previously updated the Nix expression to copy these in at
build time, but then found we cannot do that because it creates a cyclic
dependency which causes flox to be rebuilt when floxpkgs changes, which
causes a publish of the built flox package back to floxpkgs, and then
the cycle continues.

Until we find a better way to prevent this (probably by putting the
catalog data on a different branch) we'll just have to keep these files
in sync by hand - hopefully they won't change often.