fix: present multiple packages for upgrade (#548)
## Current Behavior

The `flox list --json` feature was correctly calculating all of the
available package upgrades, but the "diff-catalogs.jq" script was
dropping all but the first upgrade when presenting its results.

## Proposed Changes

This patch introduces the recursive call which causes it to walk the
entire list of available adds/upgrades/removals when presenting its
results. Fixes FLX-150.

## Checks

- [x] All commits in this Pull Request are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)
and Verified by Github.
- [x] I have accepted the flox [Contributor License
Agreement](../blob/main/.github/CLA.md) by adding my Git/Github details
in a row at the end of the
[CONTRIBUTORS.csv](../blob/main/.github/CONTRIBUTORS.csv) file by way of
this pull request or one done previously.